Bug 1865304 - Step 3. Move `Fragment` navigation responsibility to a separate class
